fix(sdk): Remove circular dep between reactnativetracing and measurements
