+++ src/js/index.ts
@@ -34,7 +34,7 @@ export {
   getActiveSpan,
 } from '@sentry/core';
 
-import { _addTracingExtensions } from './measurements';
+import { _addTracingExtensions } from './tracing/addTracingExtensions';
 _addTracingExtensions();
 
 export {
